refactor(policy): share channel deny rule guard
